Convert Params to Decimal (and int64), model Upgrade (#84)
## Convert from double/float/uints
Doubles and floats are notoriously imprecise, and are heavily cautioned against in BlockChain usage (for anything on the chain). So we make a (very) simple Decimal type for chain usage. We could have used a library, but the code is so simple it seemed much simpler to just do it directly.

uints should not be used anywhere we have math. One mistake and we have an overrun to uint64.Max.
1. Convert all floats or doubles in Params to Decimal. It uses a coefficient and and exponent (like other similar libraries).
2. Convert uint64 to int64 when math is used on the value (so, everything but IDs.)
3. Had to convert both Params and EpochGroupData (which contans ValidationParams.
4. Lots of scattered changes to support this.

## Upgrade example
While we won't need this specific upgrade going forward, this commit demos our approach to upgrading the data in the chain:
1. Copy the unmodified .proto files for the data we are changing. Rename them with a V1 in front (or V2 if there is already a V1)
2. Change the original proto files to whatever new format we want
3. Write a custom getter for the OLD values (V1)
4. Register a change handler that gets the old values and then writes the new ones.