Ml node upgrades and Partial Upgrades (#94)
## Node Version Upgrades
### Basic Mechanism
1. In API, keep track of upcoming node_version upgrades in the config
2. Add version to the Node (and InferenceNodeConfig) objects to track version
3. Check for any upcoming upgrades each block and add them to the `NodeVersionStack` if they are new
4. At each height change, check, if it's time to change the current version in the config
5. Use the current version to find the version of the node that matches
6. For now, default to AcceptEarlierVersion, where we will default to using a versionless query for the Locked node that won't block by version. We can reconsider this later, it's a simple `bool`
#### Via standard upgrade
Along with api_binaries and binaries, allow specifying node_version. If a version is specified, it's added the the upcoming node_versions

#### Via Partial upgrade
Partial upgrade is a new object on the chain. It is a `gov` message, and must be voted on just like a regular upgrade.

If it passes, it executes and adds to the list of upcoming upgrades in the chain, which then gets queried by the APIs and put in their list of upcoming versions.

Note that partial upgrade should also support API binary only upgrades, though I have not had a chance to properly test that.