Fix Mutli-Model support (#133)
* First stab. Some tests pass, some fail, working on fixes

* Add logging, fix tests
Also a fix for all models

* Index EpochGroupData with modelId
Now we have to index and lookup EpochGroupData by both startBlockHeight and modelId, since we have one for each.
This also should make it easy to quickly lookup all the EpochGroups for an Epoch.

* Everything working!
Removed all reference to models in Participant. It's solely determined by Hardware_node, and the models reported there.
Made sure the models are counted when deciding to send an update

* Refactorings for clarity

* Use modelId instead of POC Height to track subgroups

* Fix extra joins
We weren't tearing down an extra join, so it would stick around and cause problems.