Make testermint work with warm/cold keys (#259)
* First commit for making testermint work with warm/cold keys
Overall: genesis works on a single `test` key, joins work with `file` keys, plus warm/cold
1. Add some env variables to the `join` compose file
2. Add the ability to inject passwords to the docker CLI (for file key operations)
3. Change around the init so that we launch genesis, then for each pair, we launch the chain-node, use that to setup keys/file keyring, then use that to set env variables for the api before launching
4. Update the entire API to use explicit warm or cold keys.

* Password needs to have "\n" to inject correctly
And rename to "stdin", as we'll just inject whatever.

* Send transactions directly through the CLI
No longer through the API. Works with `file` keyring-backend by injecting password

* Turn off unneeded reboot, fixup logging in BLSDKG

* We were only partially using warm key
We were still passing KEY_NAME (cold key) in the joins, and since we created the COLD key on those nodes and are sharing the keychain, we were getting away with it.
This fixes that by changing ONLY the KEY_NAME in the `api` container for joins by adding -WARM.
then it also fixes up sending the grant command, which was all wrong since it was expecting it to be a transaction (since it was in `tx` CLI namespace), but it's more of a command.

* Add handling for invalid clusters, add more careful setup for new clusters

* Fix BLSDKGSuccessTest for now

* Improve the algorithm for Dynamic Pricing verification
It was approximating the increase as always being the max, but variations in requests and other factors meant it wasn't always there.
Replace that assumption with an actual calculation (ala perBlockFactor) based on usage.

* Add delegation for blstypes.MsgSubmitGroupKeyValidationSignature˚

* Fix build- peg build env instead of latest